+++ CHANGELOG.md
@@ -5,7 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 - Added option for email notification to be sent "clean," without full logging information
-- HTML theme colors can be configured in `[theme root]\theme.conf` 
+- Added a default check for a configuration file (`deploy.sh`) in project root directory
+- HTML theme colors can be configured in `[theme root]/theme.conf` 
 ### Changed
 - Recently changed files that are .gitignored will no longer halt automated deployment
 - Running as root is no longer allowed; this can be overridden by setting `ALLOWROOT="TRUE"`